Add Voter

+++ src/Controller/Web/DeleteUser/v1/Manager.php
@@ -2,22 +2,28 @@
 
 namespace App\Controller\Web\DeleteUser\v1;
 
+use App\Application\Security\Voter\UserVoter;
+use App\Controller\Exception\AccessDeniedException;
 use App\Domain\Entity\User;
 use App\Domain\Service\UserService;
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 
 class Manager
 {
-    public function __construct(private readonly UserService $userService)
-    {
+    public function __construct(
+        private readonly UserService $userService,
+        private readonly AuthorizationCheckerInterface $authorizationChecker,
+    ) {
     }
 
-    public function deleteUserById(int $userId): bool
-    {
-        return $this->userService->removeById($userId);
-    }
-
+    /**
+     * @throws AccessDeniedException
+     */
     public function deleteUser(User $user): void
     {
+        if (!$this->authorizationChecker->isGranted(UserVoter::DELETE, $user)) {
+            throw new AccessDeniedException();
+        }
         $this->userService->remove($user);
     }
 }
